requirements.txt
python-dotenv
playsound
sounddevice
webrtcvad
numpy
scipy
openai-whisper
pyautogui
pillow 

langchain 
langchain-openai 
langchain-community